+++ git_machete/generated_docs.py
@@ -322,6 +322,12 @@ long_docs: Dict[str, str] = {
               If you want the worktree to have its own branch layout file (located under `.git/worktrees/.../machete`),
               set `git config machete.worktree.useTopLevelMacheteFile false`.
 
+           `machete.squashMergeDetection`:
+              Controls the algorithm used to detect squash merges. Possible values are:
+              * `none`: No squash merge/rebase detection.
+              * `simple`: Compares the tree state of the merge commit with the tree state of the upstream branch. This detects squash merges/rebases as long as there was not any commit on the upstream branch since the last common commit.
+              * `exact`: Compares the patch that would be applied by the merge commit with the commits that ocurred on the upstream branch since the last common commit. This detects squash merges/rebases even if there were commits on the upstream branch since the last common commit. However, it might have a performance impact as it requires listing all the commits in the upstream.
+
         <b>Environment variables:</b>
            `GIT_MACHETE_EDITOR`
               Name of the editor used by `git machete e[dit]`, example: `vim` or `nano`.